div main&mainlayout code
